requirements.txt
requests_cache==1.2.0
openmeteo_requests==1.2.0
retry_requests==2.0.0
pandas==2.2.2
numpy==1.26.3
requests==2.31.0
schedule==1.2.2
pytz==2024.1
openpyxl==3.1.3
python-dotenv==1.0.1
joblib==1.4.2
tensorflow==2.17.0
scikit-learn==1.5.1
psycopg2-binary==2.9.9